[PRFix modify error messages]

+++ cmd/oras/blob/delete.go
@@ -16,11 +16,12 @@ limitations under the License.
 package blob
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 
 	"github.com/spf13/cobra"
-	"oras.land/oras/cmd/oras/internal/errors"
+	"oras.land/oras-go/v2/errdef"
 	"oras.land/oras/cmd/oras/internal/option"
 )
 
@@ -51,9 +52,6 @@ Example - Delete blob from the insecure registry:
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			opts.targetRef = args[0]
-			if !strings.Contains(opts.targetRef, "@") {
-				return fmt.Errorf("%s: image reference not support, expecting <name@digest>", opts.targetRef)
-			}
 			return deleteBlob(opts)
 		},
 	}
@@ -69,17 +67,19 @@ func deleteBlob(opts deleteBlobOptions) (err error) {
 		return err
 	}
 
-	if repo.Reference.Reference == "" {
-		return errors.NewErrInvalidReference(repo.Reference)
+	if repo.Reference.Reference == "" || !strings.Contains(opts.targetRef, "@") {
+		return fmt.Errorf("%s: blob reference not support, expecting <name@digest>", opts.targetRef)
 	}
 
-	ref := opts.targetRef
-	desc, err := repo.Blobs().Resolve(ctx, ref)
+	desc, err := repo.Blobs().Resolve(ctx, opts.targetRef)
 	if err != nil {
+		if errors.Is(err, errdef.ErrNotFound) {
+			return fmt.Errorf("%s: the specified blob does not exist", opts.targetRef)
+		}
 		return err
 	}
 	if err = repo.Delete(ctx, desc); err != nil {
-		return err
+		return fmt.Errorf("failed to delete %s: %w", opts.targetRef, err)
 	}
 
 	fmt.Println("Deleted", opts.targetRef)
